feat: Add minimal storage engine crate (#35)
This adds a `RocksStore` type that provides a `StorageTx` allow for various
storage operations on the underlying db. Note that this a very early version,
and as such, lacks a ton features that we'll need to add in over time. Also, the
api is not set in stone, but is close to what I would want to expose to the rest
of the system.

- Every record has a primary index that acts as the key to some row. When
inserting a row, this is a list of usize values which index into the row values
themselves. Keys will be extended to included a timestamp to allow for
transactional semantics.

- Rows are inserted as a "primary row internal value". The other type of
"internal value" is a tombstone, which indicates a record deletion. "Internal
values" will be expanded to encompass things like write intents for interactive
and distributed transactions.

- Everything (keys and values) is currently serialized using bincode. As such,
there are very light guarantees about ordering during a scan.

- This does not map super well to lemur's source traits, and there will need to
be some modifications to both the traits and the api presented here.

- Very inefficient. This is very much a "let's get something working".

- Minimal thought has been put into recovery. Rocks has a WAL, which is fine for
now, but will be lacking when raft is added on top.