feat: Implement lemur data source for rocks store (#44)
* feat: Implement lemur data source for rocks store

* Wire up cli to allow for specify memory or rocksdb for storage

* Add rocksdb cursor

This unfortunately wasn't as clean as I wanted it to be. Ideally this could be
turnined into an asynchronous stream, but the lifetime of the RocksDB iterator
is tied to the lifetime of the StorageTx reference. Getting this to work would
require some extensive trait redesign to allow for this to work, so going to
hold off on that for now.

Ideally we could do something like:

```
    async fn scan(
        &self,
        table: &RelationKey,
        filter: Option<ScalarExpr>,
    ) -> Result<Option<DataFrameStream>> {
        let iter = self
            .cursor(table.clone(), &[], filter)?
            .map(|result| result.map_err(anyhow::Error::from));
        let stream = futures::stream::iter(iter);
        Ok(Some(Box::pin(stream)))
    }
```