feat: Implement a tweaked object storage caching layer
This will obviate everything in the `file` module. This provides direct access
to the underlying file (which is useful in async contexts) and allows for
bypassing saving to disk.

One drawback with this is every file needs to be wrapped in a mutex. Eventually
we will want to move towards something that provides true async io.

API might continue to be tweaked as work on the access layer continues.