test: Add some slt tests
I started to take some tests from the DuckDB repo since most of their tests
seems to deal with edge cases. Ideally we'll be able to get the majority of
those tests in here. DuckDB does have some behavioral differences from Postgres,
so we'll need to keep that in mind, and in cases where they differ, we should
follow what postgres does.

I commented out and/or inserted halts for tests that don't pass for us yet.

I ended up making a small change to `pgsrv` to properly propagate errors to the
client when failing to parse a sql query. Previously it was disconnecting since
we just bubbled it up all the up to the server. Once we're further along, we'll
need to review how we want to propagate errors here since we don't want parse
failures or general bad user input causing the connection to close.