feat: Add option to output logs in json (#257)
This will make logs a bit more readable in the cloud console. There might be
mismatches between how the `tracing` lib structures logs, and what GCP expected.
We'll have to figure those out, but this should get us started.

This does not get us Cloud Trace or OpenTelemetry tracing. We will want to have
support for those eventually, but it's not an immediate need.

Closes https://github.com/GlareDB/glaredb/issues/246.