chore: Move `AccessRuntime` to `access` crate
This moves `AccessRuntime` from `sqlexec` to `access`.

The rationale for this change is this runtime will be shared across more crates
than just `sqlexec`. In particular, the catalog will be its own crate, and I
planned on making use of the same runtime that we're using during sql execution
for the catalog. There would be a circular reference between the `sqlexec` and
`catalog` crates if `AccessRuntime` remained in `sqlexec` (since `sqlexec` will
be importing `catalog`).

The `catalog` crate will make use of the same `AccessRuntime` that's being used
in `sqlexec` for data persistence.