chore: Add stale.yml
Issues marked stale after 21 days, closed after an additional 7.

Rationale for this is we probably shouldn't have issue be open for a month with
no activity at this stage of product development. If an issue is open for longer
than this with no activity, it either doesn't matter, or the issue isn't
actionable in its current state.