chore: Rework background worker to be multi-threaded
Previously ran on a single thread. Since we skipped missed ticks, this could
result in some workers not actually being ran. This change makes it so that
every background job has its own (tokio) thread with independent interval
checking.

This also introduces the `BackgroundJob` trait to act as an interface to for all
background jobs. This made stubbing out a dummy job for test pretty simple. This
can be expanded in the future to accomadate background jobs for other parts of
the system.