feat: Add discrete bootstrap steps (#350)
* feat: Add discrete bootstrap steps

Pretty much code-defined migrations.

This adds a "bootstrap" table to track which steps have been ran, only running
steps that have yet to be run. This removes the need for explicitly check if
things exist before making changes to the system.

Eventually these steps will be hardened and will be treated as a list of
append-only migrations. But for now, there will likely be changes to these
steps (which may lead to broken databases).