fix: error when reporting usage on deleted db
Summary
---

When a db is deleted in the cloud product, there is eventual consistency
between the record being deleted and the underlying k8s/cloud resources
being deleted. Therefore, there is a moment in time whereby glaredb may
be running `report_usage` on a healthy pod for a database that doesn't
exist on Cloud.

To cover this scenario, we silently accept statuses of 404 when pushing
usage.

Resolves: GlareDB/cloud#393