feat: Add full views for 'schemata', 'tables', and 'columns' (#534)
... for `information_schema`.

Also added a few postgres system tables as well.

A lot of these return null, which is fine for now. As we add more views, we can
revisit whether or not what we're returning is correct and complete. I'm
concerned about structure so that these can be used and tested.