fix: Fix allowing duplicate names (#592)
* fix: Fix allowing duplicate names

Attempting to add connections (and other objects) with the same name should now
always error.

The fix was the order in which we appended builtin objects. In short, we we're
overwriting the `schema_objects` map for the "public" schema. No tests caught
this for two reasons:

1. Builtins were added later.
2. No tests were added that intermingled a user object with a builtin object.

In addition to 1, we were also incorrectly persisting builtin objects. I've made
that change, and added a small comment to the storage proto, so that this is
documented in multiple places (in addition to what's in `builtin.rs`).

* Check that we don't load builtins