fix: Unbreak CTEs (#776)
This bug was from my change with planning where we pull out table references by
visiting the ast. A CTE name is a valid table reference, so we would then try to
find a table provider with that name, which would then fail.

The change is skip returning an error in the case of not finding a table
provider for a reference. There will be edge cases to consider if we want to do
fancier table resolution, but this should work for now.