chore: Remove storing columns directly on external tables
Closes #857

Instead these should/will be stored on the table options. Currently this just
has columns for the internal tables, I didn't feel like we needed to keep the
columns for the other data sources since we have the virtual catalog now.

In the future, if we think storing column info is worth it, we can lazily load
and store the column info using the virtual catalog.