fix: Change logic when determining when to exit worker
Previously we were getting the strong count from the cached state reference. But
this reference is changed on every call to `fetch`, and so was inaccurate.

This change uses the `version_hint` reference instead. This reference is stable,
and so will provide an accurate count of sessions relying on the worker. I would
consider this a "quick fix", but I also don't think it's terrible.

Closes #984