chore: Log error if failed to find catalog (#1173)
This happens rarely, and corresponds to redeploys of metastore. After some time,
it does fix itself.

The fix is to fallback to loading the catalog from storage if it's not in the
map. But before doing that, I think we should look at how we want to evict
catalogs, because implementing that would solve this issue too.

The logging is really there to tell us how often this happens.