chore: Swap bracket character in prompt
"〉" is actually pretty uncommon for fonts. We could go with "⟩" since it's
typically preferred over "〉", but it still lacks support in some popular
monospace fonts like Source Code Pro and Fira Mono.

As someone who likes to make everything "just right", I wouldn't like for a tool
that I'm using to use a glyph not support by my font. And so a basic ">" seems
like the most straightforward pick for the prompt character.