fix: Use python print for `show` method (#1511)
* fix: Use python print for `show` method

This is a fix for `show` not outputting anything in Google Colab. Stdout in
colab gets routed through a few different layers, and just flushing wasn't doing
anything. Using python's builtin `print` was just the easiest path for this, and
since this isn't performance sensitive, I think it's reasonable to do.

Closes https://github.com/GlareDB/glaredb/issues/1470

* util