feat: Update catalog with rpc
This change makes it so that the local session will pull the catalog after it
detects the query was a DDL query.

```
GlareDB (v0.4.0)
Type \help for help.
Using in-memory catalog
> \open http:localhost:6542
Connected to Cloud deployment: unknown
> create table hello7 (a int);
Table created
> insert into hello7 select 1 as a;
Inserted 1 row
> insert into hello7 select 2 as a;
Inserted 1 row
> select * from hello7;
┌────────────┐
│ $operation │
│ ──         │
│ Utf8       │
╞════════════╡
│ 1          │
│ 2          │
└────────────┘
```

Other changes:

- Reverts some of the execution stream/result stuff. It made working with the
results pretty clumsy.
- Adds a quick-fix around querying system tables by wrapping it in a local hint.
This was useful for debugging.
- Reinits the datafusion context on client attach. This was done to create a
catalog mutator that didn't have a metastore client. Without this change, the
catalog would inadvertently revert back to the local session's catalog since the
session will attempt to update the catalog from the local cache with the client.

Bugs:

- The above example, the column name is incorrect. I'm not sure why yet, but
plan on looking into it in a follow up.
- The memory store not working is very inconvenient for testing. The above
example was done by specifying a data dir.