fix: Ask for password each time we want to ignore it (pgsrv).
`node-postgres` always sends an auth message when trying to connect
even when the backend replies with `AuthenticationOk`. This hangs the
handler when trying to connect using `psql` since the handler keeps on
waiting for a new message which it would never receive.

We simply ask the frontend to send a password when `--ignore-pg-auth`
flag is passed (since that's what `node-postgres` does), ignore the
password message and move on like nothing happened. However, this does
require that there's only 1 password message.

Fixes #1663

Signed-off-by: Vaibhav <vrongmeal@gmail.com>