fix: Be resilient to empty string values in mongo
Left a comment in code, but we'll want to determine if this is something we
should keep around. Currently mongodb will read empty values in a csv as an
empty string and not null, and so we were widening types from Float64 to Utf8
when encountering an empty string in tests, leading to a flake.

This change just assumes we can't infer anything from an empty string, which
will let us widen to the correct type when merging schemas.