fix: Allow returning query response through pgsrv with empty column
This specifically fixes the case where we query a csv file with an unnamed
column. Previously we were explicity checking the column name was empty, and now
we don't.

This has the consequence of allowing a user to select the empty column with
`select "" from ...`, but I guess that's expected.

Closes https://github.com/GlareDB/glaredb/issues/1750