chore: Remove temp catalog, session vars (mostly) from remote session
Goal: Try to make the remote session as stateless as possible to allow for it to
execute any plan for a specific database.

There's still some need for `SessionVars` when creating the datafusion
runtime/context, but I just used the default for now. It's likely that we'll
just be provided a well-typed struct for session configuration soonish.