wip(for discussion): Undo log
This is the contents of the stash I referenced here: https://github.com/GlareDB/glaredb/blob/main/crates/sqlexec/src/planner/physical_plan/create_table.rs#L154-L166

This is the very beginning of an undo log for database catalogs. I don't expect
this PR to be merged, nor do I think this is the final design, it's being pushed
more for the discussion aspect around catalog diffs and/or transaction logging
for catalogs.