fix: Don't parse version-hint as a number for iceberg tables (#2281)
Read the first line of the `version-hint.text` and take it as it is to
generate the metadata file name. Omit parsing it as a number since there
are cases when it doesn't fit in an `i32` type. Instead of parsing it as
an `i64` or `i128` we should simply skip it since such a version number
seems like a bug and can happen again but shouldn't stop us from reading
the table.

Fixes #2277

---------

Signed-off-by: Vaibhav <vrongmeal@gmail.com>