chore: Add .clang-format
Adds a `.clang-format` file for configuring `clang-format` (I use it for
auto-formatting protobufs).

My motivation for adding this in is by default clang-format will collapse a
protobuf message with a single field and no comments into a single line.

For example, it will format this:

```protobuf
message Schema {
  repeated Field columns = 1;
}
```

Into this:

```protobuf
message Schema { repeated Field columns = 1; }
```

I don't like that way of formatting, and the only reason we had it in the code
base was because I was using a unconfigured clang-format.

Unfortunately I was not able to find a specific option for this (there's a ton,
so I might've just missed it:
https://clang.llvm.org/docs/ClangFormatStyleOptions.html), so instead I just
enabled the Google style. This seems to do mostly what I want.