add commit-sha to cli (#2411)
closes https://github.com/GlareDB/glaredb/issues/2280

now it'll show a nice git version instead of just the pkg version. 
```sh
> ./glaredb
GlareDB (v0.8.0-11-g167702f)
```

you can read more about how the {git_version} is generated
[here](https://docs.rs/built/0.7.1/built/#git2).

TLDR; it uses the head (if its a tag, it uses just that, otherwise it
uses a concatonation of the last tag + the commit)


## note for development purposes. 

It does cache this in your `target` dir, so if it's not showing the
current commit, it's likely that this generated file is cached.