fix: fixedsizelist in delta (#2607)
fixes the issue of delta converting fsl -> list when writing delta
tables.

This PR adds some metadata containing the original field if it was
downcast when writing to delta, then compares those fields. If they are
different, it applies a projection from the delta field to the original
field.

If we need further performance, we'll likely have to read and write
directly instead of using delta's arrow/datafusion logic as they have
[no intention](https://github.com/delta-io/delta-rs/issues/2162) on
supporting FixedSizeList.