chore: add postgres dbt model to dbt tests (#2703)
I know there are issues here, but sharing for early review:
- Changes the expected test outcome for the view materialization in the
existing dbt test from fail to pass
- Adds a test confirming that dbt will work with an external postgres
database. This uses the actual external demo postgres instance that we
mention in our docs, so is probably brittle. Ideally, we would spin
something up (Docker?) for the purpose of the test.
- Overall, this could also probably use some refactoring, but I _think_
that that will take some pytest magic which may make the tests less
legible.

---------

Co-authored-by: tycho garen <garen@tychoish.com>