fix: Retry writing lease when error code 429 (too many requests).
This should fix the issue. We are retrying here as suggested by GCS
since writing the same object again and again is rate limited to once
per second.

Another thing that might help is writing to a new "temp" object every
time (instead of using the same one for the same process).

Fixes: https://github.com/GlareDB/cloud/issues/2746

Signed-off-by: Vaibhav <vrongmeal@gmail.com>