fix: respect bigquery limit (#2754)
Closes #716

----

I found this as I was looking through the issues earlier today. 

One concern with this (and limit pushdowns in general,) is that we
could be _under_ reporting data in these cases:

Imagine we only partially push down a predicate but we completly push
down the limit. If we get `n` rows back from the data source, but then
further filter it down, then we've returned fewer than the target
number of rows. Right? 

This isn't new, and I'm not sure we _shouldn't_ do this because of
that, and also I could imagine datafusion passing in different limits
than what the user specified.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Vaibhav Rabber <vrongmeal@gmail.com>